Middle/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
yfinance
tensorflow
pandas
numpy
